refactor: useElementWidth 및 useHasScrollElement 훅 추가
